Organize DrawableMap namespace
